Title in imperative tone and under 50 characters
Body of the commit message. This is only necessary for commits that need more explanation.
Like the title, you should explain her the 'why' and 'what' of the commit, not the 'how'.
Remember that commits should be small and independent.
No need to resolve many issues with one commit.
Each line of the body should be under 72 characters.

The optional footer is for issue trackers like GitHub